Performance before:

```
====================== Flix Compiler Throughput ======================

Throughput (best): 33.902 lines/sec (with 24 threads.)

  min:  7.255, max: 33.902, avg: 28.358, median: 31.888

  The highest throughput was in iteration: 8 (out of 10).
  The ratio between the best and worst iteration was: 4,7x.

Finished 10 iterations on 65.953 lines of code in 28 seconds.
```

Performance after:

```
====================== Flix Compiler Throughput ======================

Throughput (best): 35.641 lines/sec (with 24 threads.)

  min:  7.418, max: 35.641, avg: 29.879, median: 34.401

  The highest throughput was in iteration: 9 (out of 10).
  The ratio between the best and worst iteration was: 4,8x.

Finished 10 iterations on 65.953 lines of code in 27 seconds.
```