@magnus-madsen 2023-10-22
Just for fun, could one of you: @mlutze , @JonathanStarup, @jaschdoc, @sockmaster27 see if they can reproduce the speedup using `--Xbenchmark-phases`