@jaschdoc 2023-10-22
Using `--Xbenchmark-phases` (you used `--Xbenchmark-throughput`?)

Before (master)

```
====================== Flix Compiler Phases ======================

Mean runtime per phase (low to high):
  EntryPoint                       0.0 ms (00.0%)
  Deriver                          2.6 ms (00.0%)
  Tailrec                          9.8 ms (00.2%)
  VarNumbering                    15.4 ms (00.2%)
  Reducer                         25.5 ms (00.4%)
  Safety                          26.1 ms (00.4%)
  Reader                          26.5 ms (00.4%)
  LambdaLift                      37.0 ms (00.6%)
  PatternExhaustiveness           39.6 ms (00.6%)
  Stratifier                      43.7 ms (00.7%)
  EarlyTreeShaker                 45.8 ms (00.7%)
  LateTreeShaker                  74.2 ms (01.1%)
  Simplifier                      81.3 ms (01.3%)
  ClosureConv                     89.9 ms (01.4%)
  Lowering                        96.2 ms (01.5%)
  Kinder                         105.2 ms (01.6%)
  Lexer                          136.9 ms (02.1%)
  Namer                          145.8 ms (02.3%)
  Redundancy                     150.1 ms (02.3%)
  MonomorphEnums                 151.0 ms (02.3%)
  Weeder                         161.2 ms (02.5%)
  Instances                      192.6 ms (03.0%)
  Optimizer                      294.6 ms (04.6%)
  Resolver                       315.7 ms (04.9%)
  Parser                         460.7 ms (07.1%)
  Monomorph                      640.1 ms (09.9%)
  JvmBackend                     880.3 ms (13.6%)
  Typer                          2225.7 ms (34.4%)

Finished 10 iterations on 65,953 lines of code in 55 seconds.
```

After (virtual threads)

```
====================== Flix Compiler Phases ======================

Mean runtime per phase (low to high):
  EntryPoint                       0.0 ms (00.0%)
  Deriver                          2.6 ms (00.0%)
  Tailrec                         11.0 ms (00.2%)
  Reader                          25.4 ms (00.4%)
  Safety                          27.4 ms (00.4%)
  VarNumbering                    32.3 ms (00.5%)
  LambdaLift                      32.9 ms (00.5%)
  Reducer                         40.1 ms (00.6%)
  EarlyTreeShaker                 42.4 ms (00.7%)
  PatternExhaustiveness           46.0 ms (00.7%)
  Stratifier                      49.2 ms (00.8%)
  ClosureConv                     63.7 ms (01.0%)
  LateTreeShaker                  81.4 ms (01.3%)
  Lowering                       101.8 ms (01.6%)
  Lexer                          102.0 ms (01.6%)
  Simplifier                     108.8 ms (01.7%)
  Kinder                         115.3 ms (01.8%)
  Redundancy                     145.5 ms (02.2%)
  MonomorphEnums                 147.8 ms (02.3%)
  Namer                          159.1 ms (02.4%)
  Weeder                         182.9 ms (02.8%)
  Instances                      211.8 ms (03.3%)
  Optimizer                      291.1 ms (04.5%)
  Resolver                       314.1 ms (04.8%)
  Parser                         502.7 ms (07.7%)
  Monomorph                      593.3 ms (09.1%)
  JvmBackend                     874.2 ms (13.4%)
  Typer                          2201.2 ms (33.8%)

Finished 10 iterations on 65,953 lines of code in 56 seconds.
```