@magnus-madsen 2023-10-22
Oops. I meant --Xbenchmark-throughput. Can you update your comment?