@jaschdoc 2023-10-22
Before (master)

```
====================== Flix Compiler Throughput ======================

Throughput (best): 16,118 lines/sec (with 8 threads.)

  min:  5,115, max: 16,118, avg: 13,769, median: 15,688

  The highest throughput was in iteration: 6 (out of 10).
  The ratio between the best and worst iteration was: 3.2x.

Finished 10 iterations on 65,953 lines of code in 54 seconds.
```

After (virtual threads)

```
====================== Flix Compiler Throughput ======================

Throughput (best): 16,322 lines/sec (with 8 threads.)

  min:  5,139, max: 16,322, avg: 13,403, median: 14,664

  The highest throughput was in iteration: 7 (out of 10).
  The ratio between the best and worst iteration was: 3.2x.

Finished 10 iterations on 65,953 lines of code in 55 seconds.
```